requirements.txt
fastapi
uvicorn[standard]
pdfplumber
python-dotenv
python-multipart
google-generativeai